requirements.txt
torch==2.8.0
torchvision==0.23.0
numpy==2.0.2
h5py==3.14.0
PyYAML==6.0.2
pyrallis==0.3.1
wandb==0.21.3
torchinfo==1.8.0
tqdm==4.67.1
psutil==7.0.0
matplotlib==3.10.6
imageio==2.37.0
imageio-ffmpeg==0.6.0
Pillow==11.3.0
einops==0.8.1
einx==0.3.0
diffusers==0.11.1
transformers==4.41.2
huggingface_hub==0.23.4
tensorboardX==2.6.4
vector-quantize-pytorch==1.22.16
termcolor==3.1.0
gymnasium==1.2.0
Shimmy==2.0.0
dm-control==1.0.15
mujoco==3.3.5
egl_probe==1.0.2
robosuite==1.5.1
